Update openfaas-loki adapter so it clones git
Issue has been opened to address this issue:
https://github.com/LucasRoesler/openfaas-loki/issues/32

While waiting for a response from the author,
I've updated the `functions` role so that the helm
chart is cloned to the `/tmp` folder.

This change also introduces a new role `gitlab` to push a gitlab runner
Gitlab runner can be used to set up a gitlab CI runner to run CI jobs.